moved git service into commons
